Fixed: #81549 - New push/run scripts added.
This is a rollup of all commits for the branch script_updates.

Removing old script files.

Added new scritps for pushing files and running commands.

Setting perms on /etc/HPCCSystems to be owned by hpcc.

Adding in script files to install.

Updating install to make sure perms are right and fixing copyright utf8 logo.

Fixed: #81549 - Fixed issue presented by link dereferencing if /etc/HPCCSystems is a symlink.

Fixed: #81549 - Changing if syntax that has issues on shell and older versions
of bash.

Updated hpcc-run.sh.in to have Richards fix for Bug: #82053.

Fixed: 82093 - Removed force of changing permissions as it is not needed and breaks sub dirs in the directory.

Removed invalid lines from the buildCmd function

Signed-off-by: Philip Schwartz <philip.schwartz@lexisnexis.com>