BUG:83072 Remove legacy code from the grammar
Remove some experimental features from the ecl grammar which are no
longer supported.  It also rounds out the implementation of inline
function definitions (which are primarily used by the FUNCTIONMACRO
command).  For detailed documentation see the bug report.

It corrects the encoding of a couple of example files, adds a couple
of test cases, and fixes a problem with the make file when no warnings
are generated.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>