BUG: 80885 Set and show Compiled state for WUActionCompile WU
Existing code sets and displays a Completed state for a workunit even
if a user wants to just do compile using Compile buttons in ECL IDE.
The user may not be able to distinguish between the WU's they have
submitted and those they have only compiled.

I change the eclccserver code to set the Compiled state if the action
is WUActionCompile. ECLWatch dispays the state in WU Details page. To
avoid people thinking it is 'stuck' in compiled state, I add an Action
field on the top of the state field. The Action is read from WU xml.
For this situation, the Action shows Compile.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>