BUG:83665 method to allow clients to change dali throttle limits
Default is 10 concurrent requests, and throttling by 1000ms.
Which means that if there are >10 active requests from a client to dali, it
will delay up to max 1s (if one becomes free it will delay less)

Some clients may want to raise this level if expected to have more
concurrent activities. Esp is an example, where under some configurations,
blocking if >10 actives sessions may be undesirable.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>