FIX #71 - Compiler warnings from modern g++
Fix a number of compiler warnings related to use of ... with format
strings that don't (or might not) match the supplied parameters.

Most were most likely harmless but a couple looked like they would
cause unintended results.

g++ does not like use of printf etc with a variable for the format
string parameter (because it will behave unpredicatably if the
string passed to it contains a %). I don't like using appendf with a
single constant string parameter (because it's inefficient).

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>