FIX #90 Removed error with component type check
Remove the change of bash's internal file seperator (IFS) in hpcc_common.

Change the usage of which to type --path as which can throw errors when a
search executable is not found on the path.

Remove regex pruning of the IP list because it placed an unneeded strain
on the usage of IFS to break down the list leaving extra spaces.

Change method of skipping dali ip from above regex pruning to an if check
if the dali ip matches the current iteration ip.

Corrected formatting and verified that tabs are not present.

Signed-off-by: Philip Schwartz <philip.schwartz@lexisnexis.com>