BUG:83808 Speed up regex and add algorithm option
Reenables some old code in 64bit that was previously disabled because
it used too much stack in 32bit.  It also allows a HINT to be specified
on a PARSE to allow the stack or heap implementation to be selected.
Add HINT(ALGORITHM('stack')) to select the recursive stack based version
Add HINT(ALGORITHM('heap')) to select the heap based version.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>