FIX: 84151 - install-cluster.sh fails if run from home directory
If run from the home directory of the admin user that you name,
install-cluster.sh would copy the tgz file it created over itself, install it,
then delete it. It will then fail to install on the other nodes.

Change the location where the tgz file is created to /tmp/ to avoid this
Also delete the tgz file when done.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>