FIX #137 - Visual Studio 2010 build issues
Avoid issues in VS2010 where a custom build step uses a batch file.

The bison and flex build commands we use may be implemented as batch
files that set up environment variables to enable bison to locate any
files it needs before passing parameters on to the executable. In such
cases the custom build steps generated for VS2010 by CMake need to use
CALL to call the nested batch file otherwise only the first custom build
step of any project will be executed.

Also, some errors in the dependencies for generated files caused VS2010
to attempt to rebuild the same file several times in parallel. This
could lead to internal errors in hidl or unnecessary rebuilds. These
are fixed in this patch.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>