Bug: 81434 Release DFU WU lock as soon as possible
In the existing FileSpray service, the DFU WU lock is released
at the end of a service method (Spray, Despray, etc). The lock
may be held longer than it is needed. One or 2 extra functions
may be executed after a submit() or commit() is called.

A {} pair is added in each service method. The { is at one line
before a DFU WU is created. The } is one line after a submit()
or commit() is called. The new code releases DFU WU lock as soon
as the DFU WU is out of the scope. The extra functions which do
not need the DFU WU are out of the scope.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>