Bug: 83861 Add more debugging/postmortem information to ESP log
Existing ESP log contains a log line to trace ESP processing time
for each HTTP request. The code is modified to add more information
into the log line if there is an exception or the request takes a
long time to process. The information includes user (if available),
IP, and HTTP/SOAP request.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>