FIX #162 Unrecognized interface resolved as 0.0.0.0
If the interface named in environment.conf does not exist, the code
was supposed to fall into using hostname when resolving the '.' net
address. However due to a flaw in the logic it will actually use
0.0.0.0, and various components will fail to start as a result.

This patch splits the 'test' and 'retrieve' functionality into
separate functions to make them easier to follow, and adds the ability
to use a wildcard in the interface specification. When searching for
an interface ip, loopback interfaces are always searched AFTER
non-loopback ones.

With this fix in place, the most appropriate value for interface= in
environment.conf is *

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>