Bug: 84908 Get correct cluster name to retrieve 'prefix' for cluster
When a user opens the ECLWatch Activity page, the ESP log contains
a warning message: "WARNING: Failed to get environment information".
The message was created when ECLWatch tried to retrieve the 'prefix'
for each target cluster. The old code retrieves the prefix using
a process name, which is incorrect. The code is modified to convert
the process name to target cluster name and use the cluster name to
retrieve the prefix.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>