Roxie not processing continuation data properly
When reading from disk rather than in-memory (preload), data from
flat or CSV files from fileparts other than the first may appear
truncated if there is more data to be returned from slaves than can
be returned in a single chunk.

There was some confusion about what BufferedDirectReader::tell()
should return - it was returning a file position in the entire
underlying file, where the corresponding 'seek' to continue reading
was assuming the position was relative to those parts of the file
managed by this node.

With this patch, tell() refers to the position in the stream rather
than in the complete logical file, which is consistent with how the
code in ccdactivities expects it to behave.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>