FIX: #216 Upload/download menu not work for Windows
The Upload/download menu link does not work for windows deployment.
When the menu is clicked, the page displays an error message:
"Dropzone is not found...". That is because an OS flag is not set
if OS is windows. The dropzone path cannot be built correctly
without the OS flag. This fix sets the OS flag based on the
environment.xml.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>