Issue #215 Fix EclWatch File Download on multi-node system
The problem only happens on multi-node system (ECLWatch and dropzone
are not in the same machine). When a user wants to download a file
using ECLWatch Upload/download File menu, the file download should be
done by clicking the file on the Upload/download page. The old code
does not work because network IP is not used as a part of the input
of createIFile(). The fix creates a RemoteFilename object using both
IP and file path/name. Then, the RemoteFilename object is used as the
input of createIFile(). Now, the fie download works.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>