Add option to use git describe to generate BUILD_TAG
Add an option to use git describe to generate the value for BUILD_TAG.
This should help differentiate internal/test versions more easily.
Default the option off for now, since it is not reliable - it will not
be regenerated if the git status changes after cmake is run.

However anyone wanting to generate a RPM/DEB that they
intend to share or load to a shared cluster may want to enable it.

We could move the generation of the build_config.h (or the part of it
that generates #define BUILD_TAG) into a custom target that is executed
on every make (but which only modifies the file if the contents is
different.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>