WIP #174 Fix problems with count on temporary file
Counts on temporary files have problems because the filename is
implicitly mangled with the wuid.  This wasn't being taken into
account.

Add code to output some more flags from the code generator, and code
in hthor to use those flags.

Note, this is likely to still require a similar fix in thor.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>