Bug: 84908 Fix Warning log for EclWatch Activity page
The existing code tries to retrieve Prefix information for a physical
cluster using getPrefixName(). The getPrefixName() logs a warning if
the information cannot be found. But, the information does not exist
since the prefix is not associated with a physical cluster. The
warning should not be there. This fix removes getPrefixName() which
is used for retrieving the Prefix information for a physical cluster.
The prefix field of TpCluster is set to "".

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>