Bug: 85010 Fix "Disk full" error while getting system topology
A "CFileIO::write, Disk full" error occurs when ECLWatch tries to
create and display a system topology (clusters or servers) page.
The write() should not be called when the page is created and
displayed. The problem is caused by a group of debug code. This
fix removes the debug code.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>