Bug:#85204 Removes exponential searches from eclcc
Replaces a few situations where an expression was being walked as
a tree rather than a DAG.  This caused eclcc to take a very long time
to process  complicated grouping conditions.  The new code ensures
children of a common graph node are only visited once.

Not high priority fix since it is a very unusual condition to hit
(this was hit in unusual auto-generated code.)

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>