EclAgent through pipe behaving erratically
Fixes gh-269. Due to a race condition between the reading and writing threads
at either side of the through PIPE activity, erratic results could be returned.
These would include spurious reports of pipe program failing with error code
254, or incomplete results.

This patch abandons the original unreliable EclAgent implementation of the PIPE
activity family and uses instead the implementation copied from Roxie, adapted
for the eclagent execution environment. An error in the Roxie implementation
(which also would have affected the original hthor code) concerning the handling
of REPEAT,GROUP pipes is also corrected. The pipe8.ecl test from the regression
suite is corrected to run properly on linux. Finally, a flag is added to allow the
return code from the pipe program to be ignored (for easier testing of the cases
where empty results are returned by allowing us to use grep).

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>