ISSUE #80: Support custom result display via workunit resources Edit Add support for defining custom result views to the resource manifest.
Result views are implemented as XSLT stylesheets and are declared in the
manifest using the following syntax.

<Manifest>
<Resource type="XSLT" filename="xslt_dependencies/included.xslt"/>
<Resource type="XSLT" filename="result.xslt" name="displayResults"/>
<Views>
<Results type="XSLT" resource="displayResults" name="display results" />
</Views>
</Manifest>

If the xslt includes other local xslt files they must be added as
addtional resources.

If the xslt includes server side xslt files the include statement can use
a path starting with "/esp/".

Signed-off-by: Anthony Fishbeck Anthony.Fishbeck@lexisnexis.com