FIX: #225 Fix ConfigMgr Wizard to allow overlap of thor master and slave
Fix ConfigMgr Wizard to allow overlap of thor master and slave node. A
prompt will be displayed if the number of ip addresses entered is more than
10 and user specifies the same number of thor slaves as the number of ip
addresses.

The message of the prompt is as follows

"As the number of Thor slave nodes requested is equal to the number
of ip addresses given, there will be an overlap of Thor master node
and a Thor slave node. This is not recommended in an environment
with more than 10 nodes.

Do you want to continue?"

If the user answers yes to this question, environment is generated.
Otherwise, user is given a chance to edit the number of thor slave nodes.

Signed-off-by: Sridhar Meda <sridhar.meda@lexisnexis.com>