Fix gh-204 - Roxie fails ASSERT_SINGLE_THREADED test
There was a check in CMessageCollator::add_package that was intended to
confirm that the mapping data structure was accessed on a single thread
only and therefore did not require protecting by a critSec.

If udpInlineCollation was enabled (it defaults off) then the check was
triggered, since the function is called by more than one thread.

Remove the check and add a CriticalBlock to prevent problems if two
threadas access this funtion simultaneously.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>