Bug: 86235 Check user input in Local Path using dropzone path
When a file is sprayed, the Local Path field is pre-populated using
a dropzone directory. A user has to input a file under that directory
or select a file using the Choose File button. The existing code
verifies the file selection by comparing the Local Path field with
pre-populated value of the dropzone directory. If they are the same,
an error message will be shown. But, when the spray request comes from
a modified DFU WU, the Local Path field is pre-populated using the
full path of the file input by the user. The existing code does not
handle that. This fix changes the code to remember the value of the
dropzone directory and compare the Local Path field with the value of
dropzone directory.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>