Bug: 86227 Avoid truncating large file when uploading file
A large file may be truncated when the file is uploading and
the Upload Now button is clicked again. This fix disables the
Upload Now button while a file is uploading. This fix also adds
a warning message about not navigating away from the uploading
page. A temporary name is used for the uploaded file before the
uploading is finished.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>