gh-357 - Roxie not using the CSV separator info from dali
Roxie was not picking up the SEPARATOR, QUOTE and TERMINATOR
settings from dali when reading a CSV file.

Add a queryProperties method to the IResolvedFile interface,
and use it to track the properties from an IDistributedFile.
While the full IDistributedFile info is not serialized to slaves
in the delayed resolution cases, the properties need to be.

Fixes gh-357. See also bugzilla issue 83402.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>