REGRESSSION: typo in code to ensure full reading could core
A typo in the code for ensureing that read/pread read the full amount even
when interrupted by signals could result in a heap corruption, as the byte
before the buffer might get overwritten.

Code was checking a 64-bit unsigned -1 value against a 32-bit one, and thus
comparing 0xffffffffffffffff versus 0x00000000ffffffff and not getting a match

In fact neither type was right - the return value is ssize_t (which is signed)

Fixes gh-254 (again). No builds containing this bug were released externally.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>