More roxie changes from cppcheck
Roxie has to initialize activity member variables in start() for child
queries to work correctly, and therefore frequently did not bother to do
so in constructor. cppcheck flags thes, and it is probably right that it
is poor practice.

This fix resolves the majority of cppcheck-reported issues in Roxie,
excluding those in non-Roxie header files (to be done later), some false
positives, and a couple that need more thought.

No issues fixed in this commit are believed to have caused any problems
in practice.

Signed-off-by: Richard Chapman <rchapman@precision390.(none)>