Fixes gh-368 - Assignment in test condition in wujobq
Not clear what the symptoms of this would be - I don't know how/when the
code reaches this branch rather than the qdata-based one.

Fix the incorrect assigment and also a missing increment of i, which
together would have meant that requesting any item other than the head
would have returned the last item rather than the one requested.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>