Fix gh-415 - problems for adding files to existing superfile
Two problems: (1) the "OK" button is not enabled when a user
inputs the superfile name using mouse's copy/patse; (2) the
'Add to an existing superfile' flag is not sent to server for
Firefox and Chrome. To fix problem (1), both oninput attribute
and onmouseout attribute are added for the 'Superfile Name'
input field. The oninput attribute is for FF and Chrome. It
does not work for IE. I tested other attributes, such as
onchange, onpaste, onmouseup, and onmouseout. Only the last
one works. To fix problem (2), the id attribute is added for
the ExistingFile input.

I also removed the Cancel button since its function has been
included inside the "Go Back" button.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>