Fix gh-424 Roxie fails to load rest of queryset if any query fails
If any query in a QuerySet fails to load, Roxie would not load the rest of
the QuerySet. The error reporting in such cases could also be improved.

The particular case that was causing trouble was a query with dllname="".
This should be reported better by Roxie and should never have been allowed
to be published by workunit.cpp

Add a try-catch block in Roxie to allow the remainder of the QuerySet to
process if any query or alias fails to load. Add code to workunit.cpp to
detect empty dllname when publishing a query.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>