StringMapOf::replace(MAPPING) incorrect
Spotted while reviewing other code.

inline bool remove(MAPPING * mem) in StringMapOf and MapOf templates is not
correct - it should be calling removeExact rather than remove, and would be
a lot clearer if it was itself called removeExact. It looks like a bit of
cut-and-paste with insufficient edits...

Rename the function removeExact, and fix it to call base::removeExact. The
prior definition has been made private with no implementation, so that any
accidental usage of it will be caught. There does not appear to be any such
usage at present.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>