Assignment in condition that only causes issues for badly formed http
Fixes gh-367.

While this is a very simple change with a low level of risk, the risk of not
taking this change is also extremely low.  I think the assignment has has
been in place since at least 2002.

For all well formatted HTTP headers, this assignment has no effect.  The value
of the character at that position would already be ':', only in the case in which
an HTTP header line didn't contain a ':' delimiter, the "value" pointer would
have been set to an position past the end of the string.

Signed-off-by: Anthony Fishbeck <Anthony.Fishbeck@lexisnexis.com>