gh-450 Add support for roxie listening on a queue
To make Roxie-on-demand more efficient, and to make Roxie more like other
components and thus simplify the middleware, Roxie now has the ability to
listen for workunits on a job queue in the same way that hthor and thor do.

If a port value of 0 is specified for a roxie server to listen on, it will
create a threadpool that listens on a job queue rather than on a socket.
When workunits are added to the queue a temporary query factory is created,
executed, and destroyed.

Slaves will do a lazy create of the required IQueryFactory if a request is
received for a workunit that is not loaded. This is added to a cache so that
subsequent slave requests on the same channel are efficient. A broadcast
message at the end of the job allows the IQueryFactory to be removed from
the cache and released.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>