bug: #84683 Fixes SEH with very large data values
If a *fixed length* string or data value was larger than 100k the
temporary was created in a different way.  Unfortunately the memory
was never allocated which caused the program to crash.
Many instances of large fields would work as expected.

Uses a template class to avoid problems passing parameters to
constructors for class members.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>