gh-478 - CLinuxDirectoryIterator not threadsafe
CLinuxDirectoryIterator is using the non-threadsafe function readdir rather
than the threadsafe version readdir_r. This appears to be causing segfaults
in readdir64 when more than one thread does a directory lookup at the same
time.

Recent changes to xref mean that this is much more likely to occur in
dafilesrv than it was previously.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>