Adds an option to eclcc to compile a definition
Instead of specifying a filename to eclcc the -main option allows an
ECL style reference of a definition within the known sources/repository
to be used instead.

e.g., eclcc -I mysources -main a.b.c
may be equivalent to
eclcc -I mysources mysources/a/b/c.ecl

If the source file is an archive -main can be used to select a
definition within that archive as the definition to compile.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>