Xref change holding reference could crash
The gh-352 change, unintentionally introduced a reference to the endpoint
when should have been a copy, consequently those endpoints could become
invalid and sequently cause a core on access. This would tend to happen
when a client already had a few previous connections to a dafilesrv,
but where the original connection (CRemoteFile) had been destroyed along
with it's endpoint.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>