Fix: gh-552 Usage graph not shown on Chrome
From a Thor usage page, a usage graph should be displayed after a
user clicks the Get Usage Graph button. In Chrome browser, existing
ECLWatch page does display a graph frame, but, does not display the
bars for workunits inside the frame. That is because the existing
code displays the bars using the same functions as used for IE
browser and the functions do not work for Chrome. This fix displays
the bars using the functions which work for IE, FF, and Chrome.

When a mouse is moved on the top of the bar, a popup window is shown
to display information about the workunit. This fix checks whether
window.createPopup is supported by a brwoser or not. If yes (IE),
window.createPopup is used to display the popup window. If not
(FF and Chrome), a YUI panel is used for the popup window. The YUI
panel is not used for IE because, for IE, the panel cannot be
displayed on the top of the SVG Usage Graph. BTW: a YUI popup is not
used because HTML cannot be used inside a YUI popup window.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>