gh-587 WsDfu should use scope to check user's file access
WsDfu displays file access permission inside the Logical File
Details page. Existing WsDfu code checks the permission using the
logical file name as the input. In LDAPSecurity dll, the input and
the permission are cached to save time for the same input. Since
the file permissions are file scope based, WsDfu should use the
file scope as the input to check the permission for a better
performance. This fix changes the input from logical file name to
file scope. In addition, the code for the file permission check is
moved to a new function "getUserFilePermission" in order to
improve code readability.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>