gh-591 Set replicate flag based on cluster setting
This fix improves usability in ECLwatch Spray/copy page. The
request originally comes from the Bugzilla:
https://bugzilla.seisint.com/show_bug.cgi?id=88205.

For the thors, if the "replicate outputs" is false, it should
not replicate and not display the 'replicate" checkbox; if the
'replicate outputs" is true, it should display the checkbox and
set it 'checked' as default. For hthor, display the checkbox and
set it 'checked' as default.

In addition, for copy page, the onChangeGroup() is moved from
onChangeMachine() to onload. The onChangeMachine() is called
whenever Source Drop Zone selection is changed. The onChangeGroup()
should only be called when Destination cluster is changed or the
Copy page is loaded.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>