gh-533/531 Fix wrong display format in Results window
Right now, ECLWatch results window displays values centred
vertically rather than top justified. The text and values are
displayed in a serif font rather than sans serif font. So, the
results become very difficult to read. That is because web
style files are not loaded by an incorrect 'if condition'.
Those style files should be included unless the web page is
used to build Excel output. The fix checks whether the output
format is 'xls' or not. If not, the style files will be linked.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>