Fix gh-640 Fix roxie diskWrite::resolve failure
In CRoxieServerDiskWriteActivity::resolve, a call is made to the workunit
to query the cluster name. The cluster name "roxie" from the workunit
does not resolve in subsequent calls in CRoxieWriteHandler to add the
cluster name.
Fix the problem by using the current cluster name, store in the
roxieName string variable. This allows the cluster to be resolved and
files to be created.

Signed-off-by: William Whitehead <william.whitehead@lexisnexis.com>