gh-644 Change button names to avoid misleading
In ECLWatch View Log page, people have been confused by the name
of the Submit button and Reversely button. The Submit button
shows the log lines in the order of the earliest first. The
Reversely button displays most recent line first. This fix
changes the Submit button to be EarliestFirst button. Also
change the Reversely button to be LatestFirst button.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>