Fixes #464 Display wizard dialog when user is done with Summary View
As suggested in the issue, display wizard dialog when user presses ok button
in Summary View. The wizard dialog should be displayed only if no other file
is already open in the advanced mode. Since the message is not being
displayed, the message not being displayed the second time around is not
an issue anymore.

Bugzilla 81119 is similar to this issue, so marking that as a duplicate.

Signed-off-by: Sridhar Meda <sridhar.meda@lexisnexis.com>