FIX gh-656 Workunits stuck in 'compiling' state
Queries sumbitted to roxie via ecl ide were left in the 'compiling'
state. This was due to a combination of bugs:

1. eclccserver was only checking the agent queue, not the roxie queue
2. eclccserver was not setting the state to failed or compiled if agent
   queue was empty
3. If action was not specified, a separate bug meant that the agent queue
   check was accidentally bypassed and the submit then succeeded.

This patch ensures that eclccserver sets the state appropriately on all
paths, and eliminates the distinction between agent and roxie queues as
a given target cluster can only have one or the other.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>