Fix gh-652 Roxie dali caching
Roxie needs to be able to save away sufficient information retrieved from dali so
that if can successfully start up even if dali is down.

The code to achieve this had a number of issues in it, notably related to when the
information was refreshed if roxie DID connect to dali, and how information fetched
from xpaths that contained filters was cached. This patch fixes those issues.

Also, add the ability to lock/unlock dali access. When locked, only the cached info
will be used.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>