Fix gh-666 Roxie locks up when trying to overwrite a file
The recent refactoring of Roxie file write code is locking up when
trying to delete an existing logical file (when ,OVERWRITE specified),
as detach() is called while a lock is still held on the file because
of the properties.

Add a clear() of properties field in CResolvedFile::remove to ensure
that the detach can succeed.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>