Fix gh-685 thorlib.group() gives internal error in eclagent
thorlib.group() (deprecated) is giving an internal error in eclagent. It's
deprecated because it is not always possible to give an unambiguous answer,
since there may be several thor clusters associated with a job with different
nodegroup names.

However, in the interests of harmony with legacy users, we should make the
effort to return an answer where an unambiguous one is available.

If all thors associated with the wu use the same nodegroup, or if there are
no thors associated, return an appropriate value. Otherwise, fail.

See also bugzilla ticket 89275

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>