Fixes #692 Fix if condition for New and New Range to act on "Computer" type only.
This is a side effect of the fix for bugzilla bug 81208. New and New range
for Computer acts differently from the New for the other types but all
types were being handled by that piece of code causing the menu item not
to work. Fixed the if condition to act on Computer type only.

Signed-off-by: Sridhar Meda <sridhar.meda@lexisnexis.com>