Fix gh-695 Spill file fails creation
Spill files created via diskWriteActivity (as opposed to spillActivity)
have the TDXjobtemp flag set (as opposed to TDXtemporary). Because of
this the code incorrectly mangles the filename and leaves it in a form
(with ::) that cannot be created.
Fix the problem by checking both of these bits, when resolving filename
and before publishing.

Signed-off-by: William Whitehead <william.whitehead@lexisnexis.com>