Package file resolution issues
A couple of issues in the package file code for resolving superfiles:

1. Negative memory leaks in two places
2. Using the query name as the package name doesn;t work very well with the
   standard alias scheme.

As well as fixing the mermory corruption issues, add a "queries=wildcardspec"
field to a package that allows it to specify  for what queries this package
should be used. These are checked in order, so queries=* will replace the
previous 'default' package name.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>