New ECL Command Line tool
New command line tool to provide uniform way of working with ecl,
workunits, and queries (published workunits).

Initial implementation supports 'Deploy' and 'Publish' commands.

ecl deploy [options][<archive>|<eclfile>|<so>|<dll>]

   Options:
      --cluster=<cluster>  cluster to associate workunit with
      --name=<name>        workunit job name

      --server=<ip>        ip of server running ecl services (eclwatch)
      --port=<port>        ecl services port

Deploy takes a local ECL archive, shared object, or dll and loads it onto
the HPCC.  If successful the Deploy command results in a workunit on the
HPCC in the 'compiled' state.

ecl publish [options][<wuid>]

   Options:
      --cluster=<cluster>  cluster to publish workunit to
                           (defaults to cluster defined inside workunit)
      --name=<name>        query name to use for published workunit
      --wuid=<wuid>        workunit id to publish
      --activate           activates query when published
      --server=<ip>        ip of server running ecl services (eclwatch)
      --port=<port>        ecl services port

Publish registers an existing workunit as a query in a queryset.

Signed-off-by: Anthony Fishbeck <Anthony.Fishbeck@lexisnexis.com>