Should indicate a 'diabled in this version' feature
Fix gh-723: ECLWatch may need to show a page indicating a
'diabled in this version' feature. For example, right now, the
"Search Roxie Files" could be a 'diabled in this version' feature.
When the link "Search Roxie Files" is clicked, ECLWatch should
shows "This feature is disabled in this version". This fix adds
the message page into WsSMC, which may be used for other similar
features.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>