Fix gh-683 - Reduce length verify timeout
Sasha if idle periodically verified it's connection with Dali, with a very
long (5 mins) timeout. If Dali was down at the time Sasha was stopped and
Sasha was verifying the connection at this point, there was no way to cancel
it (as stuck re-connecting with this timeout in the MP layer).
See separate issue gh-773.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>