Fix gh-803 User-Agent not set if no LDAP enabled
ECLwatch does not display ReLogin menu correctly when no LDAP
is enabled. That is because that ECLWatch function depends on
ESP to provide browser information. The information is retrieved
from HTTP User-Agent header. The existing code reads the
User-Agent header inside an authentication related function
which is called if no LDAP is enabled. This fix moves the
User-Agent code out of the authentication function.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>