Fix gh-843 Thor Usage Graph show NaN%
Using the latest code, I can see the Thor Usage Graph using
the Get Usage Graph button. But, some 'Business %' and
'non-Business %' show NaN%. One bug is found inside the new
code added recently and I fix it now.

Another change is this fix is to pass the target cluster name
to the Uasge page. When the Get Thor Queue button is clicked,
the target cluster name is used to retrieve the queue length
for that target cluster.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>