Fix issue with status page and shared thors
A shared thor, i.e. one that is in >1 cluster and listening to >1 queue
caused the activity page some confusion. Resulting in not displaying some
running jobs.
This is really a workaround until activiy page is restructued to more
acturately reflects the topology.
The workaround has Thor store the target cluster of workunit in the running
Status, which is then retreieved by the ws_smcService to tie up which target
the running workunit belongs to.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>